PR feedback from Jon
