+++ src/identity-providers/routes.ts
@@ -1,14 +1,12 @@
 import type { RouteDef } from "../route-config";
-import {
-  IdentityProviderCreateRoute,
-  IdentityProviderRoute,
-} from "./routes/IdentityProvider";
+import { IdentityProviderRoute } from "./routes/IdentityProvider";
 import { IdentityProviderKeycloakOidcRoute } from "./routes/IdentityProviderKeycloakOidc";
 import { IdentityProviderOidcRoute } from "./routes/IdentityProviderOidc";
 import { IdentityProviderSamlRoute } from "./routes/IdentityProviderSaml";
 import { IdentityProvidersRoute } from "./routes/IdentityProviders";
 import { IdentityProviderAddMapperRoute } from "./routes/AddMapper";
 import { IdentityProviderEditMapperRoute } from "./routes/EditMapper";
+import { IdentityProviderCreateRoute } from "./routes/IdentityProviderCreate";
 
 const routes: RouteDef[] = [
   IdentityProviderAddMapperRoute,
